+++ constants.py
@@ -15,6 +15,9 @@ materialLibrary = adsk.core.Application.get().materialLibraries.load(ASSETS_FOLD
 minimumGemstoneSize = 0.05
 gemstoneOverlapMergeThreshold = 0.01
 
+channelInset = 0.005
+channelJunctionSphereScale = 1.1
+
 patternAlongPathDefaultSpacingCm = 0.5
 patternAlongPathPlacementOnSurfaceIndex = 0
 patternAlongPathPlacementOnCurveIndex = 1